chore(API): add validation for user sign up create user sign up validation middleware function [Finishes#167161258]
