+++ server/controllers/userController.js
@@ -8,28 +8,6 @@ const { user_secret } = process.env;
 
 class UserController {
   static userSignUp(req, res) {
-    const schema = Joi.object().keys({
-      email: Joi.string().email().required(),
-      first_name: Joi.string().min(3).required(),
-      last_name: Joi.string().min(3).required(),
-      password: Joi.string().alphanum().min(6).required(),
-      phoneNumber: Joi.number().required(),
-      address: Joi.string().required()
-    });
-    const { error: validationErrors } = Joi.validate(req.body, schema, {
-      abortEarly: false
-    });
-    if (validationErrors) {
-      const error = [];
-      const { details: errors } = validationErrors;
-      errors.forEach(element => {
-        error.push(element.message.split('"').join(''));
-      });
-      return res.status(400).json({
-        status: res.statusCode,
-        error: error
-      });
-    }
     const { email, first_name, last_name, password, phoneNumber, address } = req.body;
     const token = jwt.sign({ email: email }, user_secret);
     const newUser = { token, id: users.length + 1, email, first_name, last_name, password, phoneNumber, address, is_admin: false }
