+++ server/middlewares/validation.js
@@ -18,6 +18,7 @@ function validator(req, res, schema, next) {
   next();
 }
 class ValidationMiddleware {
+  // Create new Property Validation
   static createPropertyValidation(req, res, next) {
     const schema = Joi.object().keys({
       owner: Joi.number().integer().required(),
@@ -31,6 +32,7 @@ class ValidationMiddleware {
   }
 
   static updatePropertyValidation(req, res, next) {
+    // Update Property details validation
     const schema = Joi.object().keys({
       owner: Joi.number().integer().min(1),
       price: Joi.number().min(3),
@@ -38,6 +40,19 @@ class ValidationMiddleware {
       city: Joi.string().min(3),
       address: Joi.string().min(3),
       type: Joi.string().min(3)
+    });
+    validator(req, res, schema, next);
+  }
+
+  // User Sign Up validation
+  static userSignUpValidation(req, res, next) {
+    const schema = Joi.object().keys({
+      email: Joi.string().email().required(),
+      first_name: Joi.string().min(3).required(),
+      last_name: Joi.string().min(3).required(),
+      password: Joi.string().alphanum().min(6).required(),
+      phoneNumber: Joi.number().required(),
+      address: Joi.string().required()
     });
     validator(req, res, schema, next);
   }
